Cache data as interface
