+++ libraries/apollo-normalized-cache-api-incubating/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/api/OperationCacheExtensions.kt
@@ -1,6 +1,7 @@
 package com.apollographql.apollo3.cache.normalized.api
 
 import com.apollographql.apollo3.annotations.ApolloExperimental
+import com.apollographql.apollo3.api.CompositeAdapter
 import com.apollographql.apollo3.api.CustomScalarAdapters
 import com.apollographql.apollo3.api.Executable
 import com.apollographql.apollo3.api.Operation
@@ -10,7 +11,6 @@ import com.apollographql.apollo3.api.json.MapJsonWriter
 import com.apollographql.apollo3.api.toJson
 import com.apollographql.apollo3.api.variables
 import com.apollographql.apollo3.cache.normalized.api.internal.CacheBatchReader
-import com.apollographql.apollo3.cache.normalized.api.internal.CacheDataTransformer
 import com.apollographql.apollo3.cache.normalized.api.internal.Normalizer
 
 fun <D : Operation.Data> Operation<D>.normalize(
@@ -106,7 +106,7 @@ private fun <D : Executable.Data> Executable<D>.readInternal(
     cache: ReadOnlyNormalizedCache,
     cacheResolver: Any,
     cacheHeaders: CacheHeaders,
-): CacheDataTransformer<D> {
+): CacheData {
   return CacheBatchReader(
       cache = cache,
       cacheHeaders = cacheHeaders,
@@ -115,11 +115,21 @@ private fun <D : Executable.Data> Executable<D>.readInternal(
       rootKey = cacheKey.key,
       rootSelections = rootField().selections,
       rootTypename = rootField().type.rawType().name
-  ).collectData(adapter())
+  ).collectData()
 }
 
 fun Collection<Record>?.dependentKeys(): Set<String> {
   return this?.flatMap {
     it.fieldKeys()
   }?.toSet() ?: emptySet()
+}
+
+fun <D: Executable.Data> CacheData.toData(
+    adapter: CompositeAdapter<D>,
+    customScalarAdapters: CustomScalarAdapters,
+): D {
+  val reader = MapJsonReader(
+      root = toMap(),
+  )
+  return adapter.fromJson(reader, customScalarAdapters)
 }
